fix: handle offset views correctly (#40)
<!-- Provide additional contextual information about the code changes
below this line, then remove this line. -->

It's possible for a consumer to pass in a `Uin8Array` that is pointing
at some offset of the underlying `ArrayBuffer`. This is fine for most of
the code in this library, except when we create our own `DataView` on
the underlying `ArrayBuffer`. Previously we were creating that
`DataView` at the `0` index of the underlying `ArrayBuffer`, now we
create it at `Uint8Array.offset` instead to ensure we do not read the
wrong indexes.